+++ src/main/java/edu/kis/vh/stacks/Stack.java
@@ -1,7 +1,7 @@
 package edu.kis.vh.stacks;
 
-import edu.kis.vh.stacks.list.StackArray;
-import edu.kis.vh.stacks.list.StackList;
+import edu.kis.vh.stacks.implementation.StackArray;
+import edu.kis.vh.stacks.implementation.StackList;
 
 /**
  * @author nik
