+++ src/main/java/edu/kis/vh/stacks/implementation/StackList.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.stacks.list;
+package edu.kis.vh.stacks.implementation;
 
 import edu.kis.vh.stacks.IStack;
